python_code_for_streamlit/requirements.txt
snowflake-connector-python[pandas], 
plotly, 
pandas